pyproject.toml
[project]
name = "taggerservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.6.0",
    "anyio==3.7.1",
    "certifi==2023.7.22",
    "charset-normalizer==3.3.0",
    "click==8.1.7",
    "clip-interrogator==0.6.0",
    "colorama==0.4.6",
    "exceptiongroup==1.1.3",
    "fastapi==0.103.2",
    "filelock==3.12.4",
    "fsspec==2023.9.2",
    "h11==0.14.0",
    "huggingface-hub==0.31.2",
    "idna==3.4",
    "jinja2==3.1.2",
    "markupsafe==2.1.2",
    "mpmath==1.3.0",
    "networkx==3.0",
    "numpy==1.26.0",
    "onnxruntime>=1.22.0",
    "opencv-python>=4.11.0.86",
    "packaging==23.2",
    "pandas>=2.2.3",
    "pillow==10.0.1",
    "py-consul>=1.6.0",
    "pydantic==2.4.2",
    "pydantic-core==2.10.1",
    "python-multipart==0.0.6",
    "pyyaml==6.0.1",
    "requests==2.31.0",
    "sniffio==1.3.0",
    "starlette==0.27.0",
    "sympy==1.12",
    "tqdm==4.66.1",
    "typing-extensions==4.8.0",
    "urllib3==2.0.6",
    "uvicorn==0.23.2",
]
